Update heatmap color encoding
Take all of the `dl_frac` values for a location
across a year (binned by the groups visible
on the heatmap) and find the mean value, as well
as the min and max. Encode the mean value to look
more gray, with the min and max being more colorful.

The idea is to highlight the distribution and show
you where most tests are and where no tests are.

Refs #297